Add update extension
