+++ data/tasks.txt
@@ -1,3 +1,3 @@
-T,false,minecraft speedrun
+T,true,minecraft speedrun
 T,false,myhomework
-T,false,a,2025-01-02
+T,false,something
