+++ src/main/java/micromanager/Launcher.java
@@ -1,0 +1,13 @@
+package micromanager;
+
+import micromanager.graphics.Main;
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
