+++ src/main/java/micromanager/commands/ListCommand.java
@@ -1,0 +1,26 @@
+package micromanager.commands;
+
+import micromanager.storage.Storage;
+import micromanager.storage.TaskList;
+
+/**
+ * ListCommand class represents a command to list all tasks in the task list.
+ * It extends the Command class and provides methods to execute the command.
+ */
+public class ListCommand extends Command {
+    /**
+     * Executes the list command by displaying all tasks in the task list.
+     * If the task list is empty, it displays a message indicating no tasks.
+     *
+     * @param taskList The list of tasks to display.
+     * @param storage  The storage handler (not used in this command).
+     */
+    @Override
+    public String execute(TaskList taskList, Storage storage) {
+        if (taskList.size() == 0) {
+            return "Yay! You have no tasks!";
+        } else {
+            return "Here are the tasks in your list:\n" + taskList;
+        }
+    }
+}
