+++ text-ui-test/runtest.bat
@@ -14,7 +14,7 @@ IF ERRORLEVEL 1 (
 )
 REM no error here, errorlevel == 0
 
-REM run the program, feed duke.commands from input.txt file and redirect the output to the ACTUAL.TXT
+REM run the program, feed micromanager.commands from input.txt file and redirect the output to the ACTUAL.TXT
 java -classpath ..\bin Duke < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
